Update entity models and repositories for new database schema

+++ src/main/java/com/example/hangar/repository/EmpresaRepository.java
@@ -3,5 +3,5 @@ package com.example.hangar.repository;
 import com.example.hangar.model.Empresa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmpresaRepository extends JpaRepository<Empresa, Long> {
+public interface EmpresaRepository extends JpaRepository<Empresa, Integer> {
 }
