+++ src/main/java/com/example/hangar/service/EmpresaService.java
@@ -8,9 +8,9 @@ public interface EmpresaService {
 
     List<Empresa> findAll();
 
-    Empresa findById(Long id);
+    Empresa findById(Integer id);
 
     Empresa save(Empresa entity);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
