+++ src/main/java/com/example/hangar/service/HangarService.java
@@ -8,9 +8,9 @@ public interface HangarService {
 
     List<Hangar> findAll();
 
-    Hangar findById(Long id);
+    Hangar findById(Integer id);
 
     Hangar save(Hangar entity);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
