+++ src/main/java/com/example/hangar/service/PilotoService.java
@@ -8,9 +8,9 @@ public interface PilotoService {
 
     List<Piloto> findAll();
 
-    Piloto findById(Long id);
+    Piloto findById(Integer id);
 
     Piloto save(Piloto entity);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
