+++ src/main/java/com/example/hangar/service/RolService.java
@@ -8,9 +8,9 @@ public interface RolService {
 
     List<Rol> findAll();
 
-    Rol findById(Long id);
+    Rol findById(Integer id);
 
     Rol save(Rol entity);
 
-    void delete(Long id);
+    void delete(Integer id);
 }
